+++ index.js
@@ -11,6 +11,7 @@ const mv = stew.mv;
 // const log = stew.log;
 const rm = stew.rm;
 const chalk = require('chalk');
+const SilentError = require('silent-error'); // From ember-cli
 
 const defaultOptions = {
   importBootstrapTheme: false,
@@ -36,12 +37,13 @@ module.exports = {
     this.app = app;
 
     let options = extend(extend({}, defaultOptions), app.options['ember-bootstrap']);
-    if (process.env['BOOTSTRAPVERSION']) {
+    if (process.env.BOOTSTRAPVERSION) {
       // override bootstrapVersion config when environment variable is set
-      options.bootstrapVersion = parseInt(process.env['BOOTSTRAPVERSION']);
+      options.bootstrapVersion = parseInt(process.env.BOOTSTRAPVERSION);
     }
     this.bootstrapOptions = options;
 
+    this.validateDependencies();
     this.preprocessor = this.findPreprocessor();
 
     if (!this.hasPreprocessor()) {
@@ -66,28 +68,32 @@ module.exports = {
     }
   },
 
+  validateDependencies() {
+    let bowerDependencies = this.app.project.bowerDependencies()
+    if ('bootstrap' in bowerDependencies
+      || 'bootstrap-sass' in bowerDependencies) {
+      throw new SilentError('The dependencies for ember-bootstrap are outdated. Please run `ember generate ember-bootstrap` to install the missing dependencies!');
+    }
+  },
+
   findPreprocessor() {
     return supportedPreprocessors.find((name) => !!this.app.project.findAddonByName(`ember-cli-${name}`) && this.validatePreprocessor(name));
   },
 
   validatePreprocessor(name) {
-    let npmDependencies = this.app.project.dependencies();
-    let bowerDependencies = this.app.project.bowerDependencies();
+    let dependencies = this.app.project.dependencies();
     switch (name) {
       case 'sass':
-        if (!('bootstrap-sass' in npmDependencies) && this.getBootstrapVersion() === 3) {
-          this.ui.writeLine(chalk.red('Npm package "bootstrap-sass" is missing, but required for SASS support. Please run `ember generate ember-bootstrap` to install the missing dependencies!'));
-          return false;
+        if (!('bootstrap-sass' in dependencies) && this.getBootstrapVersion() === 3) {
+          throw new SilentError('Npm package "bootstrap-sass" is missing, but required for SASS support. Please run `ember generate ember-bootstrap` to install the missing dependencies!');
         }
         break;
       case 'less':
         if (this.getBootstrapVersion() === 4) {
-          this.ui.writeLine(chalk.red('There is no Less support for Bootstrap 4! Falling back to importing static CSS. Consider switching to Sass for preprocessor support!'));
-          return false;
+          throw new SilentError('There is no Less support for Bootstrap 4! Falling back to importing static CSS. Consider switching to Sass for preprocessor support!');
         }
-        if (!('bootstrap' in bowerDependencies)) {
-          this.ui.writeLine(chalk.red('Bower package "bootstrap" is missing, but required for Less support. Please run `ember generate ember-bootstrap` to install the missing dependencies!'));
-          return false;
+        if (!('bootstrap' in dependencies)) {
+          throw new SilentError('Npm package "bootstrap" is missing, but required for Less support. Please run `ember generate ember-bootstrap` to install the missing dependencies!');
         }
         break;
     }
@@ -95,21 +101,18 @@ module.exports = {
   },
 
   getBootstrapStylesPath() {
+    let nodeModulesPath = this.app.project.nodeModulesPath;
     switch (this.preprocessor) {
       case 'sass':
         if (this.getBootstrapVersion() === 4) {
-          return path.join(this.app.project.nodeModulesPath, 'bootstrap', 'scss');
+          return path.join(nodeModulesPath, 'bootstrap', 'scss');
         } else {
-          return path.join(this.app.project.nodeModulesPath, 'bootstrap-sass', 'assets', 'stylesheets');
+          return path.join(nodeModulesPath, 'bootstrap-sass', 'assets', 'stylesheets');
         }
       case 'less':
-        return path.join(this.app.bowerDirectory, 'bootstrap', 'less');
+        return path.join(nodeModulesPath, 'bootstrap', 'less');
       default:
-        if (this.getBootstrapVersion() === 4) {
-          return path.join(this.app.project.nodeModulesPath, 'bootstrap', 'dist', 'css');
-        } else {
-          return path.join(this.app.bowerDirectory, 'bootstrap', 'dist', 'css');
-        }
+        return path.join(nodeModulesPath, 'bootstrap', 'dist', 'css');
     }
   },
 
@@ -119,7 +122,7 @@ module.exports = {
         return path.join(this.app.project.nodeModulesPath, 'bootstrap-sass', 'assets', 'fonts');
       case 'less':
       default:
-        return path.join(this.app.bowerDirectory, 'bootstrap', 'fonts');
+        return path.join(this.app.project.nodeModulesPath, 'bootstrap', 'fonts');
     }
   },
 
@@ -161,25 +164,27 @@ module.exports = {
     return this.getBootstrapVersion() === 3 ? 4 : 3;
   },
 
-  treeForAddon() {
-    let tree = this._super.treeForAddon.apply(this, arguments);
+  treeForAddon(tree) {
     let bsVersion = this.getBootstrapVersion();
     let otherBsVersion = this.getOtherBootstrapVersion();
-    let componentsPath = 'modules/ember-bootstrap/components/';
+    let componentsPath = 'components/';
+
     tree = mv(tree, `${componentsPath}bs${bsVersion}/`, componentsPath);
     tree = rm(tree, `${componentsPath}bs${otherBsVersion}/**/*`);
-    return tree; // log(tree, {output: 'tree', label: 'moved'});
+
+    return this._super.treeForAddon.call(this, tree);
   },
 
-  treeForAddonTemplates() {
-    let tree = this._super.treeForAddonTemplates.apply(this, arguments);
+  treeForAddonTemplates(tree) {
     let bsVersion = this.getBootstrapVersion();
     let otherBsVersion = this.getOtherBootstrapVersion();
     let templatePath = 'components/';
+
     tree = mv(tree, `${templatePath}common/`, templatePath);
     tree = mv(tree, `${templatePath}bs${bsVersion}/`, templatePath);
     tree = rm(tree, `${templatePath}bs${otherBsVersion}/**/*`);
-    return tree; //log(tree, {output: 'tree', label: 'moved'});
+
+    return this._super.treeForAddonTemplates.call(this, tree);
   },
 
   contentFor(type, config) {